Update busybox build configuration files
The new upstream release considerably changes the configuration. This
commit is the result of running 'make oldconfig' on the configuration
files. For the udeb, the features remain mostly the same as before
except for a handful of additions; for the deb and static
configurations, most upstream defaults have been accepted and thus there
are a number of additional new features enabled.