d/rules: rework the build/install procedure
busybox build procedure is very sensitive to the environment
variables such as CFLAGS/LDFLAGS etc.  The busybox binary is
dependent on the same content of many variables, and everything
is recompiled in case any difference is found.

But upstream `install' target depends on the busybox binary.
So when we perform install, we have to provide exactly the same
environment as used during build. This is fragile.

In particular, our udeb version were built twice - because we
used `make install' to install this one, but did not have the
same environment.  The build used dh_auto_build which sets
CFLAGS/CPPFLAGS behind the scenes, but we did not use
dh_auto_install in the same way, so whole udeb busybox were
recompiled during install, but without the proper flags.

So we omit usage of `make install' to install things.  For regular
and static builds we only install a single binary anyway, we used
to run `make install' needlessly, the result was not used. But for
the udeb target, we need the binary and the links to it named after
every applet.  The list of the links is available from the upstream
build procedure in the file named busybox.links, which is created
as a separate target.  When actually installing the links, we can
use dh_link directly, - it will give us correctly setup links
(absolute vs relative according to policy) and it will create them
right in the package destination directory without using an
intermediate temporary copy.

Also the upstream build system does not work right when asked
to made more than a single target when run in parallel mode,
so we can not request two targets in one go.

The solution to all this and a few other:

1. ensure we run build with proper environment, explicitly setting
 CONFIG_EXTRA_CFLAGS & CONFIG_EXTRA_LDFLAGS as before but without
 the extra environment provided by dh_auto_build.
2. explicitly request busybox.links file during build (it does not
 heavily depend on the environment the way busybox target does)
3. in the install target, install the binary directly, and for udeb,
 additionally use dh_link to create all the needed symlinks based
 on busybox.links created above.

dh_auto_build also handled DEB_BUILD_OPTIONS=parallel=N for us.
Turn it into MAKEFLAGS=-jN explicitly.  This also gives us a
way to use d/rules -jN way for free.

Also stop installing d/tree/busybox/* twice, once explicitly from
d/rules:install, and once from d/busybox.install.