Revert "disable loadfont on hurd"
This reverts commits 99f4c2e2f79dee693aeb93c3d537ba7ee58967d1
and 5757914c8749b602a58725e3558d89ec63fa56a8.

Instead of disabling individual utils in our config, it is
better to restore the missing "depends on PLATFORM_LINUX"
in the code. Many applets misses these, it is impractical
to disable them one by one. The same thing applies to other
non-linux platforms too, eg kfreebsd.

Ref:
http://lists.busybox.net/pipermail/busybox/2011-February/074954.html
https://git.busybox.net/busybox/commit/?id=5c69ad0ecdc18cf51b312c7c82848f4438fe1c8d
http://lists.busybox.net/pipermail/busybox/2022-October/089914.html

This stuff needs to be resurrected with proper automatic detection
of PLATFORM_LINUX (should it take cross-compilation into account?)
so it is always properly defined regardless of make all*config, and
with linux-specific applets properly "depend on PLATFORM_LINUX" as
has been before the above commit.

The only problem is to write the proper automatic detection of linux
platform in kconfig terms (#ifdef __linux__ from a compiler?) and
the whole thing quickly becoming quite ugly.